Point out uneven spacing
Add a prompt to get students to reflect on the utility of manually setting levels.

Students in GEOL0202 noticed that the manually set levels were not evenly spaced, and they asked if that was "misrepresentation".  In discussion (referencing earlier plots) they easily understood why this was a helpful way to represent the data.  I suggest adding this prompt so that students have a chance to consider this explicitly.