Map projection, typos, prompt in Week 3 lesson (#10)
* Consistent Robinson projection

Switch two plots made with EqualEarth to Robinson projection, for consistency with subsequent plots.

Note I did not change the projection on a zonal selection, which has a comment that says "Note: I changed the projection" (from FM)--I assume this means that students should see the change

* Correct two small typos

* Point out uneven spacing

Add a prompt to get students to reflect on the utility of manually setting levels.

Students in GEOL0202 noticed that the manually set levels were not evenly spaced, and they asked if that was "misrepresentation".  In discussion (referencing earlier plots) they easily understood why this was a helpful way to represent the data.  I suggest adding this prompt so that students have a chance to consider this explicitly.